requirements.txt
fastapi
uvicorn
openai==0.28.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
